Ensure all code is up to date

+++ UI Menus/CustomAlgorithmMenu.cs
@@ -271,9 +271,9 @@ public class CustomAlgorithmMenu : MonoBehaviour {
         objectsToSpawnImages[4].sprite = GameManager.S.algorithmMenuCS.objectSprites[customAlgorithms[buttonNdx].objectToSpawn4];
 
         speedTexts[0].text = "<color=red>Starting object speed:</color> " + (customAlgorithms[buttonNdx].startingObjectSpeed + 1).ToString();
-        speedTexts[1].text = "<color=#FFC800>Amount to increase per level:</color> " + (customAlgorithms[buttonNdx].amountToIncreaseObjectSpeed / 10).ToString();
-        speedTexts[2].text = "<color=red>Starting spawn speed:</color> " + ((customAlgorithms[buttonNdx].startingSpawnSpeed / 10) + 0.1f).ToString();
-        speedTexts[3].text = "<color=#FFC800>Amount to decrease per level:</color> " + (customAlgorithms[buttonNdx].amountToDecreaseSpawnSpeed / 10).ToString();
+        speedTexts[1].text = "<color=#FFC800>Amount to increase per level:</color> " + ((float)customAlgorithms[buttonNdx].amountToIncreaseObjectSpeed / 10).ToString();
+        speedTexts[2].text = "<color=red>Starting spawn speed:</color> " + (((float)customAlgorithms[buttonNdx].startingSpawnSpeed / 10) + 0.1f).ToString();
+        speedTexts[3].text = "<color=#FFC800>Amount to decrease per level:</color> " + ((float)customAlgorithms[buttonNdx].amountToDecreaseSpawnSpeed / 10).ToString();
     }
 }
